YARN-9677. Make FpgaDevice and GpuDevice classes more similar to each other. Contributed by kevin su
